+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - support for distributed samplers
 - from_torcheval added to allow syncing of torcheval metrics
 - support for Python 3.14
+- added optional compile and distributed functionalities to the Model class
 
 ## [0.1.0rc1] - 2025-11-25